fastapi/requirements.txt
fastapi
pydantic[email]
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
alembic
psycopg2
strawberry-graphql[fastapi]

pytest

black
isort
flake8